- Merged ChessBoardUI and UI
