+++ src/main/java/View/ChessBoardUI.java
@@ -7,15 +7,14 @@ import UseCases.ActualValidCalculator;
 
 import javax.swing.*;
 import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
+import java.awt.event.*;
 import java.util.Objects;
 
 public class ChessBoardUI extends JPanel {
     private static LocationBitboard locationBitboard;
-    public int SQUARE_SIZE = 85;
+    static JFrame javaF=new JFrame("Chess Engine");//must be declared as static so that other class' can repaint
+    static ChessBoardUI javaUI=new ChessBoardUI();//must be declared as static so that other class' can repaint
+    public int squareSize = 64;
     private final int[] twoClicks = new int[2];
     private int numClicks = 0;
     private long highlightSquares = 0L;
@@ -23,19 +22,57 @@ public class ChessBoardUI extends JPanel {
     static JButton stalemateButton = new JButton("stalemateButton");
     static int blkPoints = 0;
     static int whtPoints = 0;
+    static boolean newGame = true; // Indicates that the game has not started yet
+    static boolean gameOver = false; // Indicates that the game is in the over and frames should not be updating
+    static String winning_msg = ""; // Set at the end of the game for the team that won
     static int border = 10; //the amount of empty space around the frame
-    private boolean isButtonConfigured = false;
+
+    static JButton pawnPromoteQueen = new JButton("stalemateButton"); // queen pawn promotion button
+    static JButton pawnPromoteKnight = new JButton("stalemateButton"); // queen pawn promotion button
+    static JButton pawnPromoteBishop = new JButton("stalemateButton"); // queen pawn promotion button
+    static JButton pawnPromoteRook = new JButton("stalemateButton"); // queen pawn promotion button
+    private final boolean isButtonConfigured = false;
 
     public static void main (String[]args){
-        JFrame frame = new JFrame();
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        frame.setLocationRelativeTo(null);
-        ChessBoardUI board = new ChessBoardUI();
-        frame.add(board);
-        frame.setMinimumSize(new Dimension(910, 725));
-        frame.setVisible(true);
+        new MainMenu();
     }
 
+    @Override
+    public void paintComponent(Graphics g) {
+        super.paintComponent(g);
+        this.setBackground(new Color(43, 45, 48));
+        this.addComponentListener(new ComponentAdapter() {
+            public void componentResized(ComponentEvent e) {
+                squareSize=(int) (Math.min(getHeight(), getWidth()-200-border)-2*border)/8;
+            }
+        });
+        drawBoard(g);
+        drawPieces(g, locationBitboard.getBitboard("whitePawn"), Images.whitePawn);
+        drawPieces(g, locationBitboard.getBitboard("whiteRook"), Images.whiteRook);
+        drawPieces(g, locationBitboard.getBitboard("whiteKnight"), Images.whiteKnight);
+        drawPieces(g, locationBitboard.getBitboard("whiteBishop"), Images.whiteBishop);
+        drawPieces(g, locationBitboard.getBitboard("whiteQueen"), Images.whiteQueen);
+        drawPieces(g, locationBitboard.getBitboard("whiteKing"), Images.whiteKing);
+        drawPieces(g, locationBitboard.getBitboard("blackPawn"), Images.blackPawn);
+        drawPieces(g, locationBitboard.getBitboard("blackRook"), Images.blackRook);
+        drawPieces(g, locationBitboard.getBitboard("blackKnight"), Images.blackKnight);
+        drawPieces(g, locationBitboard.getBitboard("blackBishop"), Images.blackBishop);
+        drawPieces(g, locationBitboard.getBitboard("blackQueen"), Images.blackQueen);
+        drawPieces(g, locationBitboard.getBitboard("blackKing"), Images.blackKing);
+        drawBorders(g);
+        drawBlackTeam(g);
+        drawWhiteTeam(g);
+        drawTimer(g);
+        drawForfeit();
+        drawStalemate();
+        drawPromotionButtons();
+        if (newGame) {
+            newGame = false;
+        }
+        drawHighlights(g);
+    }
+
+    // Holds chessboard exclusive listeners
     public ChessBoardUI() {
         // Get the current chessboard
         locationBitboard = ChessGame.getCurrentBoard();
@@ -45,12 +82,12 @@ public class ChessBoardUI extends JPanel {
                 int mouseX = e.getX();
                 int mouseY = e.getY();
                 System.out.println(locationBitboard);
-                int x = mouseX / SQUARE_SIZE;
-                int y = 7 - mouseY / SQUARE_SIZE;
+                int x = mouseX / squareSize;
+                int y = 7 - mouseY / squareSize;
                 int index = y * 8 + x;
                 boolean pieceFound = false;
 
-                if (mouseX >= 0 && mouseX < 8 * SQUARE_SIZE && mouseY >= 0 && mouseY < 8 * SQUARE_SIZE) {
+                if (mouseX >= 0 && mouseX < 8 * squareSize && mouseY >= 0 && mouseY < 8 * squareSize) {
                     for (String pieceType : locationBitboard.getAllPieces().keySet()) {
                         long bitboard = locationBitboard.getBitboard(pieceType);
                         // Want to check if there is a piece at where we clicked.
@@ -92,168 +129,278 @@ public class ChessBoardUI extends JPanel {
         });
     }
 
-    @Override
-    public void paintComponent(Graphics g) {
-        super.paintComponent(g);
-        for (int row = 0; row < 8; row++) {
-            for (int col = 0; col < 8; col++) {
-                Color color;
-                if ((row + col) % 2 == 0) {
-                    color = new Color(255, 255, 255);
-                } else {
-                    color = new Color(43, 45, 48);
-                }
-                g.setColor(color);
-                g.fillRect(col * SQUARE_SIZE, row * SQUARE_SIZE, SQUARE_SIZE, SQUARE_SIZE);
-            }
-        }
-        drawPieces(g, locationBitboard.getBitboard("whitePawn"), Images.whitePawn);
-        drawPieces(g, locationBitboard.getBitboard("whiteRook"), Images.whiteRook);
-        drawPieces(g, locationBitboard.getBitboard("whiteKnight"), Images.whiteKnight);
-        drawPieces(g, locationBitboard.getBitboard("whiteBishop"), Images.whiteBishop);
-        drawPieces(g, locationBitboard.getBitboard("whiteQueen"), Images.whiteQueen);
-        drawPieces(g, locationBitboard.getBitboard("whiteKing"), Images.whiteKing);
-        drawPieces(g, locationBitboard.getBitboard("blackPawn"), Images.blackPawn);
-        drawPieces(g, locationBitboard.getBitboard("blackRook"), Images.blackRook);
-        drawPieces(g, locationBitboard.getBitboard("blackKnight"), Images.blackKnight);
-        drawPieces(g, locationBitboard.getBitboard("blackBishop"), Images.blackBishop);
-        drawPieces(g, locationBitboard.getBitboard("blackQueen"), Images.blackQueen);
-        drawPieces(g, locationBitboard.getBitboard("blackKing"), Images.blackKing);
-        drawBorders(g);
-        drawBlackTeam(g);
-        drawWhiteTeam(g);
-        drawTimer(g);
-        drawForfeit();
-        drawStalemate();
-        g.setColor(new Color(128, 128, 128, 128));
-        int diameter = SQUARE_SIZE / 4;
-        int radius = diameter / 2;
-        for (int i = 0; i < 64; i++) {
-            if ((highlightSquares & (1L << i)) != 0) {
-                int x = i % 8;
-                int y = 7 - i / 8;
-                int centerX = x * SQUARE_SIZE + SQUARE_SIZE / 2;
-                int centerY = y * SQUARE_SIZE + SQUARE_SIZE / 2;
-                g.fillOval(centerX - radius, centerY - radius, diameter, diameter);
-            }
+    // Draws the chess board
+    public void drawBoard(Graphics g) {
+        for (int i=0;i<64;i+=2) {//draw chess board
+            g.setColor(new Color(255, 255, 255));
+            g.fillRect((int)((i%8+(i/8)%2)*squareSize)+border, (int)((i/8)*squareSize)+border, (int)squareSize, (int)squareSize);
+            g.setColor(new Color(43, 45, 48));
+            g.fillRect((int)(((i+1)%8-((i+1)/8)%2)*squareSize)+border, (int)(((i+1)/8)*squareSize)+border, (int)squareSize, (int)squareSize);
         }
     }
 
+    // Draws the chess pieces on the board
     private void drawPieces(Graphics g, long bitboard, Image pieceImage) {
         for (int i = 0; i < 64; i++) {
             if ((bitboard & (1L << i)) != 0) {
                 int x = i % 8;
                 int y = 7 - i / 8;
-                g.drawImage(pieceImage, x * SQUARE_SIZE, y * SQUARE_SIZE, SQUARE_SIZE, SQUARE_SIZE, null);
+                g.drawImage(pieceImage, x * squareSize, y * squareSize, squareSize, squareSize, null);
             }
         }
     }
 
+    // Draw the piece move highlight
+    public void drawHighlights(Graphics g) {
+        int diameter = squareSize / 4;
+        int radius = diameter / 2;
+        for (int i = 0; i < 64; i++) {
+            if ((highlightSquares & (1L << i)) != 0) {
+                int x = i % 8;
+                int y = 7 - i / 8;
+                int centerX = x * squareSize + squareSize / 2;
+                int centerY = y * squareSize + squareSize / 2;
+                g.fillOval(centerX - radius, centerY - radius, diameter, diameter);
+            }
+        }
+    }
+
+    // Draws the game borders
     public void drawBorders(Graphics g) {
         g.setColor(new Color(43, 45, 48));
-        g.fill3DRect(0, border, border, (int)(8*SQUARE_SIZE), true);
-        g.fill3DRect((int)(8*SQUARE_SIZE)+border, border, border, (int)(8*SQUARE_SIZE), true);
-        g.fill3DRect(border, 0, (int)(8*SQUARE_SIZE), border, true);
-        g.fill3DRect(border, (int)(8*SQUARE_SIZE)+border, (int)(8*SQUARE_SIZE), border, true);
+        g.fill3DRect(0, border, border, (int)(8*squareSize), true);
+        g.fill3DRect((int)(8*squareSize)+border, border, border, (int)(8*squareSize), true);
+        g.fill3DRect(border, 0, (int)(8*squareSize), border, true);
+        g.fill3DRect(border, (int)(8*squareSize)+border, (int)(8*squareSize), border, true);
+
         g.setColor(Color.BLACK);
         g.fill3DRect(0, 0, border, border, true);
-        g.fill3DRect((int)(8*SQUARE_SIZE)+border, 0, border, border, true);
-        g.fill3DRect(0, (int)(8*SQUARE_SIZE)+border, border, border, true);
-        g.fill3DRect((int)(8*SQUARE_SIZE)+border, (int)(8*SQUARE_SIZE)+border, border, border, true);
-        g.fill3DRect((int)(8*SQUARE_SIZE)+2*border+200, 0, border, border, true);
-        g.fill3DRect((int)(8*SQUARE_SIZE)+2*border+200, (int)(8*SQUARE_SIZE)+border, border, border, true);
+        g.fill3DRect((int)(8*squareSize)+border, 0, border, border, true);
+        g.fill3DRect(0, (int)(8*squareSize)+border, border, border, true);
+        g.fill3DRect((int)(8*squareSize)+border, (int)(8*squareSize)+border, border, border, true);
+        g.fill3DRect((int)(8*squareSize)+2*border+200, 0, border, border, true);
+        g.fill3DRect((int)(8*squareSize)+2*border+200, (int)(8*squareSize)+border, border, border, true);
+
         g.setColor(new Color(43, 45, 48));
-        g.fill3DRect((int)(8*SQUARE_SIZE)+2*border, 0, 200, border, true);
-        g.fill3DRect((int)(8*SQUARE_SIZE)+2*border+200, border, border, (int)(8*SQUARE_SIZE), true);
-        g.fill3DRect((int)(8*SQUARE_SIZE)+2*border, (int)(8*SQUARE_SIZE)+border, 200, border, true);
-        // Cast Graphics to Graphics2D to apply line thickness
-        Graphics2D g2d = (Graphics2D) g;
-        // Set the line thickness
-        float thickness = 18.0f;
-        g2d.setStroke(new BasicStroke(thickness));
-        // Set the color for the new borders
-        g2d.setColor(Color.BLACK);
-        // Draw the first line from (x:680, y:680) to (x:680, y:0)
-        g2d.drawLine(689, 689, 689, 0);
-        // Draw the second line from (x:0, y:680) to (x:680, y:680)
-        g2d.drawLine(0, 689, 689, 689);
+        g.fill3DRect((int)(8*squareSize)+2*border, 0, 200, border, true);
+
+        g.fill3DRect((int)(8*squareSize)+2*border+200, border, border, (int)(8*squareSize), true);
+        g.fill3DRect((int)(8*squareSize)+2*border, (int)(8*squareSize)+border, 200, border, true);
     }
 
-    // Creation of the black teams points and captured pieces
-    public void drawBlackTeam(Graphics g) {
-        g.setColor(new Color(103, 106, 110, 255));
-        g.fill3DRect((int)(8*SQUARE_SIZE)+2*border, border, 200, (int)(2.8*SQUARE_SIZE), true);
-        g.setColor(new Color(0, 0, 0, 255));
-        g.drawString("Black Team:", (int)(8*SQUARE_SIZE)+2*border + 2, border*2+ 2);
-        g.drawString(Integer.toString(blkPoints), (int)(8*SQUARE_SIZE)+2*border + 175 - border, border*2 + 2);
-    }
-
-    // Creation of the white teams points and captured pieces
+    // Creation of the white teams points and captured pieces.
     public void drawWhiteTeam(Graphics g) {
         g.setColor(new Color(52, 53, 56));
-        g.fill3DRect((int)(8*SQUARE_SIZE)+2*border, border+(int)(2.8*SQUARE_SIZE), 200, (int)(2.8*SQUARE_SIZE), true);
+        g.fill3DRect((int)(8*squareSize)+2*border, border+(int)(2.8*squareSize), 200, (int)(2.8*squareSize), true);
+
         g.setColor(new Color(220, 216, 216, 255));
-        g.drawString("White Team:", (int)(8*SQUARE_SIZE)+2*border + 2, border*2+ (int)(2.8*SQUARE_SIZE) + 2);
-        g.drawString(Integer.toString(whtPoints), (int)(8*SQUARE_SIZE)+2*border + 175 - border, border*2+ (int)(2.8*SQUARE_SIZE) + 2);
+        g.drawString("White Team:", (int)(8*squareSize)+2*border + 2, border*2+ (int)(2.8*squareSize) + 2);
+        g.drawString(Integer.toString(whtPoints), (int)(8*squareSize)+2*border + 175 - border, border*2+ (int)(2.8*squareSize) + 2);
     }
 
-    // Creation of the timer UI located at the bottom right
+    // Creation of the black teams points and captured pieces.
+    public void drawBlackTeam(Graphics g) {
+        g.setColor(new Color(103, 106, 110, 255));
+        g.fill3DRect((int)(8*squareSize)+2*border, border, 200, (int)(2.8*squareSize), true);
+
+        g.setColor(new Color(0, 0, 0, 255));
+        g.drawString("Black Team:", (int)(8*squareSize)+2*border + 2, border*2+ 2);
+        g.drawString(Integer.toString(blkPoints), (int)(8*squareSize)+2*border + 175 - border, border*2 + 2);
+    }
+
+    // Creation of the draw button.
+    public void drawStalemate() {
+        Image unscaledIcon = new ImageIcon(Objects.requireNonNull(getClass().getResource("/projectimages/DrawIcon.png"))).getImage();
+        Image stalemateIcon = unscaledIcon.getScaledInstance(50 ,(int)(squareSize) - 10, java.awt.Image.SCALE_SMOOTH);
+
+
+        stalemateButton.setText("");
+        stalemateButton.setFocusPainted(false);
+        stalemateButton.setBounds((int)(8*squareSize)+2*border + 100,border+(int)(6.15*squareSize),100, (int)(squareSize) - 10);
+        stalemateButton.setBackground(new Color(103, 106, 110));
+        stalemateButton.setIcon(new ImageIcon(stalemateIcon));
+        if (newGame) {
+            stalemateButton.addActionListener(e -> {
+                if (!gameOver) { // Makes sure the game isn't over
+                    // TODO put draw code here
+                    System.out.println("Requesting a draw");
+                    drawEndScreen("It's a draw!");
+                }
+            });
+        }
+        this.add(stalemateButton);
+    }
+
+    // Creation of the timer UI located at the bottom right.
     public void drawTimer(Graphics g) {
         // Creation of the clock Borders
         g.setColor(new Color(44, 46, 51));
-        g.fill3DRect((int)(8*SQUARE_SIZE)+2*border, border+(int)(7*SQUARE_SIZE), 200, (int)(SQUARE_SIZE), true);
+        g.fill3DRect((int)(8*squareSize)+2*border, border+(int)(7*squareSize), 200, (int)(squareSize), true);
         g.setColor(new Color(190, 178, 157));
-        g.fill3DRect((int)(8*SQUARE_SIZE)+2*border + 5, border+(int)(7*SQUARE_SIZE+5), 190, (int)(SQUARE_SIZE - 10), true);
+        g.fill3DRect((int)(8*squareSize)+2*border + 5, border+(int)(7*squareSize+5), 190, (int)(squareSize - 10), true);
+
+
         // Creation of the clock Label
         g.setColor(new Color(0, 0, 0));
-        g.setFont(new Font("TRUE TYPE_FONT", Font.BOLD, (int)(SQUARE_SIZE) - 20));
-        // Calculate string width
-        FontMetrics fm = g.getFontMetrics();
-        int stringWidth = fm.stringWidth("0:00");
-        int boxCenter = 190 / 2;
-        int stringXPosition = (int)(8*SQUARE_SIZE)+2*border + 5 + boxCenter - stringWidth / 2;
-        g.drawString("0:00", stringXPosition, border+(int)(7*SQUARE_SIZE+ (int)(SQUARE_SIZE)- 15));
+        g.setFont(new Font("TRUE TYPE_FONT", Font.BOLD, (int)(squareSize) - 20));
+        g.drawString("0:00", (int)(8*squareSize)+2*border + (int)(squareSize) - 15, border+(int)(7*squareSize+ (int)(squareSize)- 15));
     }
 
-    // Creation of the forfeit button
+    // Creation of the forfeit button.
     public void drawForfeit() {
         Image unscaledIcon = new ImageIcon(Objects.requireNonNull(getClass().getResource("/projectimages/Forfeit.png"))).getImage();
-        Image forfeitIcon = unscaledIcon.getScaledInstance(40 ,(int)(SQUARE_SIZE) - 25, java.awt.Image.SCALE_SMOOTH);
+        Image forfeitIcon = unscaledIcon.getScaledInstance(40 ,(int)(squareSize) - 25, java.awt.Image.SCALE_SMOOTH);
+
+
         forfeitButton.setText("");
         forfeitButton.setFocusPainted(false);
-        forfeitButton.setBounds((int)(8*SQUARE_SIZE)+2*border,border+(int)(6*SQUARE_SIZE),100, (int)(1*SQUARE_SIZE));
+        forfeitButton.setBounds((int)(8*squareSize)+2*border,border+(int)(6.15*squareSize),100, (int)(squareSize) - 10);
         forfeitButton.setBackground(new Color(44, 46, 51));
         forfeitButton.setIcon(new ImageIcon(forfeitIcon));
-        if (!isButtonConfigured) {
-            forfeitButton.addActionListener(new ActionListener() {
-                @Override
-                public void actionPerformed(ActionEvent e) {
-                    System.out.println("Forfeited");
+        if (newGame) {
+            forfeitButton.addActionListener(e -> {
+                if (!gameOver) { // Makes sure the game isn't over
+                    drawEndScreen(winning_msg);
                 }
             });
-            isButtonConfigured = true;
         }
-        // Add button to panel
+        this.add(stalemateButton);
         this.add(forfeitButton);
     }
 
-    // Creation of the draw button
-    public void drawStalemate() {
-        Image unscaledIcon = new ImageIcon(Objects.requireNonNull(getClass().getResource("/projectimages/DrawIcon.png"))).getImage();
-        Image stalemateIcon = unscaledIcon.getScaledInstance(50 ,(int)(1*SQUARE_SIZE) - 10, java.awt.Image.SCALE_SMOOTH);
-        stalemateButton.setText("");
-        stalemateButton.setFocusPainted(false);
-        stalemateButton.setBounds((int)(8*SQUARE_SIZE)+2*border + 100,border+(int)(6*SQUARE_SIZE),100, (int)(1*SQUARE_SIZE));
-        stalemateButton.setBackground(new Color(103, 106, 110));
-        stalemateButton.setIcon(new ImageIcon(stalemateIcon));
-        // commented out for safety
-//        stalemateButton.addActionListener(new ActionListener() {
-//            @Override
-//            public void actionPerformed(ActionEvent e) {
-//                System.out.println("Requesting Stalemate");
-//            }
-//        });
-//        this.add(stalemateButton);
+    // Draws the pawn promotion buttons
+    public void drawPromotionButtons() {
+        // Queen promotion button configurations
+        Image unscaledQIcon = new ImageIcon(Objects.requireNonNull(getClass().getResource("/projectimages/QueenPromote.png"))).getImage();
+        Image queenIcon = unscaledQIcon.getScaledInstance(40 ,(int)(squareSize / 1.75), java.awt.Image.SCALE_SMOOTH);
+
+        pawnPromoteQueen.setText("");
+        pawnPromoteQueen.setFocusPainted(false);
+        pawnPromoteQueen.setBounds((int)(8*squareSize +2.5*border),border+(int)(5*squareSize + (int)(squareSize / 1.7)),40, (int)(squareSize / 1.75));
+        pawnPromoteQueen.setIcon(new ImageIcon(queenIcon));
+
+        // Knight promotion button configurations
+        Image unscaledKIcon = new ImageIcon(Objects.requireNonNull(getClass().getResource("/projectimages/KnightPromote.png"))).getImage();
+        Image knightIcon = unscaledKIcon.getScaledInstance(40 ,(int)(squareSize / 1.75), java.awt.Image.SCALE_SMOOTH);
+
+        pawnPromoteKnight.setText("");
+        pawnPromoteKnight.setFocusPainted(false);
+        pawnPromoteKnight.setBounds((int)(8*squareSize +7.5 * border),border+(int)(5*squareSize + (int)(squareSize / 1.7)),40, (int)(squareSize / 1.75));
+        pawnPromoteKnight.setIcon(new ImageIcon(knightIcon));
+
+        // Rook promotion button configurations
+        Image unscaledRIcon = new ImageIcon(Objects.requireNonNull(getClass().getResource("/projectimages/RookPromote.png"))).getImage();
+        Image rookIcon = unscaledRIcon.getScaledInstance(40 ,(int)(squareSize / 1.75), java.awt.Image.SCALE_SMOOTH);
+
+        pawnPromoteRook.setText("");
+        pawnPromoteRook.setFocusPainted(false);
+        pawnPromoteRook.setBounds((int)(8*squareSize + 12.5 *border),border+(int)(5*squareSize + (int)(squareSize / 1.7)),40, (int)(squareSize / 1.75));
+        pawnPromoteRook.setIcon(new ImageIcon(rookIcon));
+
+        // Bishop promotion button configurations
+        Image unscaledBIcon = new ImageIcon(Objects.requireNonNull(getClass().getResource("/projectimages/BishopPromote.png"))).getImage();
+        Image bishopIcon = unscaledBIcon.getScaledInstance(40 ,(int)(squareSize / 1.75), java.awt.Image.SCALE_SMOOTH);
+
+        pawnPromoteBishop.setText("");
+        pawnPromoteBishop.setFocusPainted(false);
+        pawnPromoteBishop.setBounds((int)(8*squareSize + 17.5 *border),border+(int)(5*squareSize + (int)(squareSize / 1.7)),40, (int)(squareSize / 1.75));
+        pawnPromoteBishop.setIcon(new ImageIcon(bishopIcon));
+
+        // Action listeners for pawn promotion buttons
+        if (newGame) { // Ensures the action listeners are only added once
+            pawnPromoteQueen.addActionListener(e -> {
+                System.out.println("Promoting Queen!"); // Just a check that the button works can delete
+                //TODO Put pawn promotion function/code here on button press
+
+                // hides the buttons after a pawn promotion
+                pawnPromoteBishop.setVisible(false);
+                pawnPromoteQueen.setVisible(false);
+                pawnPromoteKnight.setVisible(false);
+                pawnPromoteRook.setVisible(false);
+            });
+            pawnPromoteKnight.addActionListener(e -> {
+                System.out.println("Promoting Knight!"); // Just a check that the button works can delete
+                //TODO Put pawn promotion function/code here on button press
+
+                // hides the buttons after a pawn promotion
+                pawnPromoteBishop.setVisible(false);
+                pawnPromoteQueen.setVisible(false);
+                pawnPromoteKnight.setVisible(false);
+                pawnPromoteRook.setVisible(false);
+            });
+            pawnPromoteRook.addActionListener(e -> {
+                System.out.println("Promoting Rook!"); // Just a check that the button works can delete
+                //TODO Put pawn promotion function/code here on button press
+
+                // hides the buttons after a pawn promotion
+                pawnPromoteBishop.setVisible(false);
+                pawnPromoteQueen.setVisible(false);
+                pawnPromoteKnight.setVisible(false);
+                pawnPromoteRook.setVisible(false);
+            });
+            pawnPromoteBishop.addActionListener(e -> {
+                System.out.println("Promoting Bishop!"); // Just a check that the button works can delete
+                //TODO Put pawn promotion function/code here on button press
+
+                // hides the buttons after a pawn promotion
+                pawnPromoteBishop.setVisible(false);
+                pawnPromoteQueen.setVisible(false);
+                pawnPromoteKnight.setVisible(false);
+                pawnPromoteRook.setVisible(false);
+            });
+        }
+
+        this.add(pawnPromoteQueen);
+        this.add(pawnPromoteKnight);
+        this.add(pawnPromoteRook);
+        this.add(pawnPromoteBishop);
+    }
+
+    // Draws the end screen for the game
+    public void drawEndScreen(String winning_msg) {
+        Graphics g = this.getGraphics();
+        g.setColor(new Color(0, 0, 0, 184));
+        g.fillRect(0, 0, (int)((8)*squareSize)+border*2, (int)((8)*squareSize)+border*2);
+
+        g.setColor(new Color(220, 216, 216));
+        g.setFont(new Font("TRUE TYPE_FONT", Font.BOLD, (int)(squareSize) - 20));
+        g.drawString(winning_msg, (int)((3)*squareSize - (squareSize + 40)), (int) squareSize * 4);
+
+        javaF.setIgnoreRepaint(true);
+        javaF.setResizable(false);
+        gameOver = true;
+
+        new GameOver();
+    }
+
+    // starts a new game instance
+    public static void newGame() {
+        // Setting initial game state
+        newGame = true;
+        gameOver = false;
+        blkPoints = 0;
+        whtPoints = 0;
+        winning_msg = "";
+        pawnPromoteQueen.setVisible(false);
+        pawnPromoteKnight.setVisible(false);
+        pawnPromoteBishop.setVisible(false);
+        pawnPromoteRook.setVisible(false);
+        ChessGame.new_game();
+        ChessBoardUI board = new ChessBoardUI();
+
+        // Setting the UI configurations
+        javaF.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        javaF.add(javaUI);
+        javaF.setSize(757, 570);
+        javaF.add(board);
+        javaF.setLocation((Toolkit.getDefaultToolkit().getScreenSize().width-javaF.getWidth())/2,
+                (Toolkit.getDefaultToolkit().getScreenSize().height-javaF.getHeight())/2);
+
+        javaF.setVisible(true);
+        javaF.setIgnoreRepaint(false);
+        javaF.setResizable(true);
+
+        javaF.repaint();
     }
 }
